Add sphinx documentation.
Change the api for the reader's aggreattion and group_by functions to
reflect that they don't actually discriminate on the passed-in
metric_name - they were just going for [0].  So change the api to make
the caller explicitly pass in the part of the query they're interested
in.

Lots of documentation changes.

Almost ready to call this 0.0.1